Rename widget to manage the VG Physical Extent Size (pesize)
