+++ src/lib/autoinstall/widgets/storage/pesize.rb
@@ -17,30 +17,39 @@
 # To contact SUSE LLC about this file by physical or electronic mail, you may
 # find current contact information at www.suse.com.
 
-require_relative "../../../test_helper"
-require_relative "./shared_examples"
-require "autoinstall/widgets/storage/vg_extent_size"
+require "yast"
+require "y2storage"
+require "autoinstall/widgets/storage/size_selector"
 
-describe Y2Autoinstallation::Widgets::Storage::VgExtentSize do
-  subject(:widget) { described_class.new }
+module Y2Autoinstallation
+  module Widgets
+    module Storage
+      # Widget to manage the LVM VG extent size
+      class Pesize < SizeSelector
+        # Constructor
+        def initialize
+          textdomain "autoinst"
+          super
+        end
 
-  include_examples "Y2Autoinstallation::Widgets::Storage::SizeSelector"
+        # @macro seeAbstractWidget
+        def label
+          # TRANSLATORS: field to enter the extent size of a new volume group
+          _("Physical Extent Size")
+        end
 
-  describe "#include_blank?" do
-    it "returns true" do
-      expect(subject.include_blank?).to eq(true)
-    end
-  end
+        def sizes
+          ["1 MiB", "2 MiB", "4 MiB", "8 MiB", "16 MiB", "32 MiB", "64 MiB"]
+        end
 
-  describe "#include_auto?" do
-    it "returns false" do
-      expect(subject.include_auto?).to eq(false)
-    end
-  end
+        def include_max?
+          false
+        end
 
-  describe "#include_max?" do
-    it "returns false" do
-      expect(subject.include_max?).to eq(false)
+        def include_auto?
+          false
+        end
+      end
     end
   end
 end
